requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=0.7.1=pypi_0
astor=0.8.0=pypi_0
atomicwrites=1.3.0=py36_1
attrs=19.1.0=py36_1
backcall=0.1.0=py36_0
blas=1.0=mkl
ca-certificates=2019.8.28=0
certifi=2019.9.11=py36_0
chardet=3.0.4=pypi_0
cloudpickle=1.2.2=py_0
colorama=0.4.1=py36_0
cudatoolkit=10.0.130=0
cudnn=7.6.0=cuda10.0_0
cycler=0.10.0=py36h009560c_0
cytoolz=0.10.0=py36he774522_0
dask-core=2.4.0=py_0
decorator=4.4.0=py36_1
dill=0.3.0=pypi_0
easydict=1.9=pypi_0
freetype=2.9.1=ha9979f8_1
future=0.17.1=pypi_0
gast=0.2.2=pypi_0
google-pasta=0.1.7=pypi_0
googleapis-common-protos=1.6.0=pypi_0
grpcio=1.21.1=pypi_0
h5py=2.9.0=pypi_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha66f8fd_1
idna=2.8=pypi_0
imageio=2.5.0=py36_0
importlib_metadata=0.19=py36_0
intel-openmp=2019.4=245
ipykernel=5.1.1=py36h39e3cac_0
ipython=7.7.0=py36h39e3cac_0
ipython_genutils=0.2.0=py36h3c5d0ee_0
jedi=0.13.3=py36_0
joblib=0.13.2=py36_0
jpeg=9b=hb83a4c4_2
jupyter_client=5.3.1=py_0
jupyter_core=4.5.0=py_0
keras-applications=1.0.8=pypi_0
keras-preprocessing=1.1.0=pypi_0
kiwisolver=1.1.0=py36ha925a31_0
libpng=1.6.37=h2a8f88b_0
libsodium=1.0.16=h9d3ae62_0
libtiff=4.0.10=hb898794_2
markdown=3.1.1=pypi_0
matplotlib=3.1.0=py36hc8f65d3_0
mkl=2019.4=245
mkl-service=2.0.2=py36he774522_0
mkl_fft=1.0.12=py36h14836fe_0
mkl_random=1.0.2=py36h343c172_0
more-itertools=7.2.0=py36_0
networkx=2.3=py_0
numpy=1.16.4=pypi_0
numpy-base=1.16.4=py36hc3f5095_0
olefile=0.46=py36_0
openssl=1.1.1d=he774522_0
packaging=19.0=py36_0
pandas=0.25.0=py36ha925a31_0
parso=0.5.0=py_0
patsy=0.5.1=py36_0
pickleshare=0.7.5=py36_0
pillow=6.1.0=py36hdc69c19_0
pip=19.1.1=py36_0
pluggy=0.12.0=py_0
promise=2.2.1=pypi_0
prompt_toolkit=2.0.9=py36_0
protobuf=3.8.0=pypi_0
psutil=5.6.3=pypi_0
py=1.8.0=py36_0
pygments=2.4.2=py_0
pyparsing=2.4.0=py_0
pyqt=5.9.2=py36h6538335_2
pytest=5.0.1=py36_0
python=3.6.8=h9f7ef89_7
python-dateutil=2.8.0=py36_0
pytz=2019.1=py_0
pywavelets=1.0.3=py36h8c2d366_1
pyzmq=18.0.0=py36ha925a31_0
qt=5.9.7=vc14h73c81de_0
requests=2.22.0=pypi_0
scikit-image=0.15.0=py36ha925a31_0
scikit-learn=0.21.2=py36h6288b17_0
scipy=1.2.1=py36h29ff71c_0
seaborn=0.9.0=py36_0
setuptools=41.0.1=py36_0
sip=4.19.8=py36h6538335_0
six=1.12.0=pypi_0
sqlite=3.28.0=he774522_0
statsmodels=0.10.1=py36h8c2d366_0
tb-nightly=1.14.0a20190603=pypi_0
tensorflow-datasets=1.0.2=pypi_0
tensorflow-gpu=2.0.0b1=pypi_0
tensorflow-metadata=0.13.0=pypi_0
termcolor=1.1.0=pypi_0
tf-estimator-nightly=1.14.0.dev2019060501=pypi_0
tk=8.6.8=hfa6e2cd_0
toolz=0.10.0=py_0
tornado=6.0.2=py36he774522_0
tqdm=4.32.2=pypi_0
traitlets=4.3.2=py36h096827d_0
urllib3=1.25.3=pypi_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_0
wcwidth=0.1.7=py36h3d5aa90_0
werkzeug=0.15.4=pypi_0
wheel=0.33.4=py36_0
wincertstore=0.2=py36h7fe50ca_0
wrapt=1.11.2=pypi_0
xz=5.2.4=h2fa13f4_4
zeromq=4.3.1=h33f27b4_3
zipp=0.5.2=py_0
zlib=1.2.11=h62dcd97_3
zstd=1.3.7=h508b16e_0
